+++ end_to_end_tests/golden-record/my_test_api_client/api/tests/unsupported_content_tests_unsupported_content_get.py
@@ -11,7 +11,7 @@ def _get_kwargs(
     *,
     client: Client,
 ) -> Dict[str, Any]:
-    url = "{}/tests/test_octet_stream".format(client.base_url)
+    url = "{}/tests/unsupported_content".format(client.base_url)
 
     headers: Dict[str, Any] = client.get_headers()
 
@@ -23,25 +23,19 @@ def _get_kwargs(
     }
 
 
-def _parse_response(*, response: httpx.Response) -> Optional[bytes]:
-    if response.status_code == 200:
-        return bytes(response.content)
-    return None
-
-
-def _build_response(*, response: httpx.Response) -> Response[bytes]:
+def _build_response(*, response: httpx.Response) -> Response[None]:
     return Response(
         status_code=response.status_code,
         content=response.content,
         headers=response.headers,
-        parsed=_parse_response(response=response),
+        parsed=None,
     )
 
 
 def sync_detailed(
     *,
     client: Client,
-) -> Response[bytes]:
+) -> Response[None]:
     kwargs = _get_kwargs(
         client=client,
     )
@@ -53,21 +47,10 @@ def sync_detailed(
     return _build_response(response=response)
 
 
-def sync(
-    *,
-    client: Client,
-) -> Optional[bytes]:
-    """  """
-
-    return sync_detailed(
-        client=client,
-    ).parsed
-
-
 async def asyncio_detailed(
     *,
     client: Client,
-) -> Response[bytes]:
+) -> Response[None]:
     kwargs = _get_kwargs(
         client=client,
     )
@@ -76,16 +59,3 @@ async def asyncio_detailed(
         response = await _client.get(**kwargs)
 
     return _build_response(response=response)
-
-
-async def asyncio(
-    *,
-    client: Client,
-) -> Optional[bytes]:
-    """  """
-
-    return (
-        await asyncio_detailed(
-            client=client,
-        )
-    ).parsed